+++ crate/operator/config.py
@@ -97,6 +97,12 @@ class Config:
     #: failed.
     CLUSTER_UPDATE_TIMEOUT = 7200
 
+    #: Time in seconds for which the operator will continue and wait to perform
+    #: a check if volume expansion has finshed successfully. Once this threshold
+    # has passed, a volume expansion is considered failed or not supported by the
+    # StorageClass.
+    EXPAND_VOLUME_TIMEOUT = 1800
+
     #: Enable several testing behaviors, such as relaxed pod anti-affinity to
     #: allow for easier testing in smaller Kubernetes clusters.
     TESTING: bool = False
@@ -312,6 +318,22 @@ class Config:
             raise ConfigurationError(
                 f"Invalid {self._prefix}CLUSTER_UPDATE_TIMEOUT="
                 f"'{update_timeout}'. Needs to be a positive integer or 0."
+            )
+
+        expand_volume_timeout = self.env(
+            "EXPAND_VOLUME_TIMEOUT", default=str(self.EXPAND_VOLUME_TIMEOUT)
+        )
+        try:
+            self.EXPAND_VOLUME_TIMEOUT = int(expand_volume_timeout)
+        except ValueError:
+            raise ConfigurationError(
+                f"Invalid {self._prefix}EXPAND_VOLUME_TIMEOUT="
+                f"'{expand_volume_timeout}'. Needs to be a positive integer or 0."
+            )
+        if self.EXPAND_VOLUME_TIMEOUT < 0:
+            raise ConfigurationError(
+                f"Invalid {self._prefix}EXPAND_VOLUME_TIMEOUT="
+                f"'{expand_volume_timeout}'. Needs to be a positive integer or 0."
             )
 
     def env(self, name: str, *, default=UNDEFINED) -> str: